rename app-provider to router
